+++ KkthnxUI/Modules/ActionBars/Bar3.lua
@@ -7,7 +7,7 @@ local _G = _G
 -- WOW API
 local CreateFrame = CreateFrame
 
---	SETUP MULTIBARLEFT AS BAR #3 BY TUKZ
+--	Setup multibarleft as bar #3 by tukz
 local ActionBar3 = CreateFrame("Frame", "Bar3Holder", RightActionBarAnchor)
 ActionBar3:SetAllPoints(RightActionBarAnchor)
 MultiBarLeft:SetParent(ActionBar3)
@@ -27,7 +27,17 @@ for i = 1, 12 do
 	end
 end
 
--- HIDE BAR
+-- Hide bar
 if C.ActionBar.RightBars < 2 then
 	ActionBar3:Hide()
+end
+
+-- Mouseover bar
+if C.ActionBar.RightBarsMouseover == true then
+	for i = 1, 12 do
+		local b = _G["MultiBarLeftButton"..i]
+		b:SetAlpha(0)
+		b:HookScript("OnEnter", function() RightBarMouseOver(1) end)
+		b:HookScript("OnLeave", function() if not HoverBind.enabled then RightBarMouseOver(0) end end)
+	end
 end